requirements.txt
srt==3.5.3
ffmpeg-python @ git+https://github.com/kkroening/ffmpeg-python@df129c7ba30aaa9ffffb81a48f53aa7253b0b4e6